+++ src/pamiq_io/vision/__init__.py
@@ -1,5 +1,5 @@
-"""Computer vision related utilities for game-io."""
+"""Computer vision related utilities for pamiq-io."""
 
-from .video_capture import OpenCVVideoCapture, VideoCapture
+from .capture import OpenCVVideoCapture, VideoCapture
 
 __all__ = ["VideoCapture", "OpenCVVideoCapture"]
