+++ src/pamiq_io/vision/capture/__init__.py
@@ -1,0 +1,6 @@
+"""Video capture module for computer vision tasks."""
+
+from .base import VideoCapture
+from .opencv import OpenCVVideoCapture
+
+__all__ = ["VideoCapture", "OpenCVVideoCapture"]
