+++ tests/vision/capture/test_base.py
@@ -1,0 +1,13 @@
+import pytest
+
+from pamiq_io.vision.capture.base import VideoCapture
+
+
+class TestVideoCapture:
+    """Tests for the VideoCapture abstract base class."""
+
+    @pytest.mark.parametrize("method_name", ["read", "width", "height", "fps"])
+    def test_abstract_methods(self, method_name):
+        """Test that VideoCapture correctly defines expected abstract
+        methods."""
+        assert method_name in VideoCapture.__abstractmethods__
